+++ packages/media-fields/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Enhancements
+
+-   Show excerpt or content snippets when attached media points to titleless posts. [#78897](https://github.com/WordPress/gutenberg/issues/78897).
+
 ### Code Quality
 
 -   Add missing `@types/react` dependency. [#78882](https://github.com/WordPress/gutenberg/pull/78882).